tracePropagationTargets should support custom domain fix
1. Extract hostname from HDX_COLLECTOR_URL and then Created a RegExp from the extracted hostname.

2. Used the extracted hostname in tracePropagationTargets.

By initializing HyperDX, use the RegExp created in the tracePropagationTargets.
By doing this, we won't need to rely on the API response to determine the server's hostname. The solution now utilizes the configuration file directly, making it more efficient and ensuring the hostname derived from the HDX_COLLECTOR_URL is always used.